requirements.txt
flask==3.0.0
google-generativeai==0.3.1
pillow==10.2.0
pymupdf==1.23.8
python-dotenv==1.0.0
requests==2.31.0 